lab7_1_outputedit

+++ lab7_1.cpp
@@ -5,7 +5,7 @@ int main() {
     int a  =  20,b = 30;
     int c = 4, output;
 
-    output = a - double(b)/c;
+    output = double(a) - double(b)/c;
     std::cout << "output: " << output << "\n";
     return 0;
 }
